update: Moved email services to services folder
